pyproject.toml
[project]
name = "health-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "deep-translator>=1.11.4",
    "google-api-python-client>=2.186.0",
    "minsearch>=0.0.7",
    "mistune>=3.1.4",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "playwright>=1.55.0",
    "pydantic>=2.12.3",
    "pydantic-ai>=1.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "streamlit>=1.51.0",
    "tqdm>=4.67.1",
    "youtube-transcript-api>=1.2.3",
]

[tool.uv.workspace]
members = [
    "service/youtube_extraction",
    "service/product_extraction",
    "service/agent",
    "service/evals",
]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
]
